fix: update stale unit tests after module renames and improve CI coverage gate
- Rename test_DownloadLoPivotCsvFilesStage to match DownloadWebTCADCsvFilesStage
- Rename test_LoPivotPlatformPipeline to match WebTCADTelemetryItemPipeline
- Update test assertions to use new class names and method signatures
- Add --cov-fail-under=80 to build.sh so pytest fails on coverage regression
- Add fail_below_threshold: true to cobertura CI action for explicit gate
